Cleanup windows warnings - remove unused vars - add a few (int) - suppress warning 4267 for call to HASH_ADD_STR with #ifdef WINDOWS - yech
Halibut still has many warnings

+++ app/i18n/stripmsg.c
@@ -153,7 +153,7 @@ void process( mode_e mode, FILE * inFile, char *fileName )
 				*cp++ = 'n';
 
 				/* read a line */
-				if (!fgets( cp, (sizeof(line)) - (cp-line), inFile )) {
+				if (!fgets( cp, (int)((sizeof(line)) - (cp-line)), inFile) ) {
 					return;
 				}
 				lineNo++;
@@ -171,7 +171,7 @@ void process( mode_e mode, FILE * inFile, char *fileName )
 			if (strchr(line, '"'))
 			{
 				printf("N_(\"");
-				len = strlen(line);
+				len = (int)(strlen(line));
 				for (i = 0; i < len; i++)
 				{
 					/* Escape double quotation marks */
